add ending enrollment state to temporary enrollment pairings
closes FOO-4111
flag = temporary_enrollments

test plan:
 • run migrations
 • enabled temporary enrollments feature flag
 • exercise Create Temporary Enrollment Pairing api
   • POST /api/v1/accounts/:account_id/temporary_enrollment_pairings
     • ending_enrollment_state
     • Allowed values: deleted, completed, inactive
 • new column should be populated with the value passed in
 • docs looks good

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/337281
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Migration-Review: Cody Cutrer <cody@instructure.com>
Product-Review: August Thornton <august@instructure.com>

+++ app/models/enrollment_state.rb
@@ -166,8 +166,14 @@ class EnrollmentState < ActiveRecord::Base
         self.state = wf_state
       elsif global_start_at < now
         if enrollment.temporary_enrollment?
-          # special case for temporary enrollments, mark them deleted once we're past the end date
-          enrollment.destroy
+          ending_enrollment_state = enrollment.temporary_enrollment_pairing&.ending_enrollment_state
+
+          case ending_enrollment_state
+          when "completed", "inactive"
+            self.state = ending_enrollment_state
+          when "deleted", nil
+            enrollment.destroy
+          end
         else
           # we've past the end date so no matter what the state was, we're "completed" now
           self.state = "completed"
